replace sqrt to sqrtf

+++ src/plain/plain.c
@@ -6,7 +6,8 @@
 #include "plain.h"
 
 
-float multiply(int size, float *first_arr, float *second_arr) {
+float multiply(int size, float *first_arr, float *second_arr)
+{
     float res = 0;
     for (int i = 0; i < size; i++) {
         res += first_arr[i] * second_arr[i];
@@ -14,15 +15,17 @@ float multiply(int size, float *first_arr, float *second_arr) {
     return res;
 }
 
-float process(int size, float *first_arr, float *second_arr) {
+float process(int size, float *first_arr, float *second_arr)
+{
     float uv = multiply(size, first_arr, second_arr);
     float uu = multiply(size, first_arr, first_arr);
     float vv = multiply(size, second_arr, second_arr);
-    float cosine = uv/(sqrt(uu) * sqrt(vv));
+    float cosine = uv / (sqrtf(uu) * sqrtf(vv));
     return cosine;
 }
 
-int plain(int binary_mod, FILE *fp1, FILE *fp2, int count, int length, float *result) {
+int plain(int binary_mod, FILE *fp1, FILE *fp2, int count, int length, float *result)
+{
     float *array1 = malloc(length);
     float *array2 = malloc(length);
     for (int k = 0; k < count; k++) {
